+++ Window/Window.d
@@ -232,7 +232,7 @@ final:
 	Surface capture(Texture.Format fmt = Texture.Format.BGRA) const {
 		Surface temp = Surface.make(this.vMode.width, this.vMode.height);
 		
-		const size_t psize = 4 * this.vMode.width * this.vMode.height;
+		const uint psize = 4 * this.vMode.width * this.vMode.height;
 		auto pixels = Memory.allocate!ubyte(psize, Mode.AutoFree);
 		
 		glReadPixels(0, 0, this.vMode.width, this.vMode.height, fmt, GL_UNSIGNED_BYTE, pixels.ptr);
